create separate folder for api service and add server url for registration
